requirements.txt
streamlit==1.28.0
chromadb>=1.4.0
langchain>=1.0.0
huggingface-hub>=0.16.4
sentence-transformers>=2.2.2
PyPDF2>=3.0.1
torch>=2.0.0
transformers>=4.57.0
